chore(keycard): keypair tests moved to a new separate test file
